src/requirements.txt
beautifulsoup4==4.8.2
pandas==1.0.3
PyYAML==5.3.1
requests==2.23
scipy==1.4.1
Scrapy==2.0.0
tqdm==4.43.0
Unidecode==1.1.1
xlrd==1.2.0
google-cloud-storage==1.29.0
gunicorn==20.0.4
